+++ src/Oro/Bundle/DataGridBundle/Tests/Behat/Context/GridContext.php
@@ -807,6 +807,7 @@ class GridContext extends OroFeatureContext implements OroPageObjectAware
      * Example: But when I filter Created At as not between "25 Jun 2015" and "30 Jun 2015"
      *
      * @When /^(?:|when )(?:|I )filter (?P<filterName>[\w\s]+) as (?P<type>(?:|between|not between)) "(?P<start>.+)" and "(?P<end>.+)"$/
+     * @When /^(?:|when )(?:|I )filter "(?P<filterName>[\w\s\/]+)" as (?P<type>(?:|between|not between)) "(?P<start>.+)" and "(?P<end>.+)"$/
      * @When /^(?:|when )(?:|I )filter (?P<filterName>[\w\s]+) as (?P<type>(?:|between|not between)) "(?P<start>.+)" and "(?P<end>.+)" in "(?P<filterGridName>[\w\s]+)"$/
      *
      * @param string $filterName